(Hopefully) fixed L13 text view getting cut off. (an iOS 7 error) Also some int -> NSInteger thingammies

+++ SecurityGame/BGKVLevelSelectViewController.h
@@ -13,6 +13,6 @@
 //Potential automatic storyboard method
 - (IBAction)goToLevelOfTag:(id)sender;
 
-+ (UIViewController *)initialVCForLevel:(int)level;
++ (UIViewController *)initialVCForLevel:(NSInteger)level;
 
 @end
